third_party/blink/web_tests/platform/generic/fast/css/shorthand-priority-expected.txt

Tests that querying the priority for a shorthand works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.style.getPropertyValue('border-bottom-style') is 'solid'
PASS e.style.getPropertyPriority('border-bottom-style') is 'important'
PASS e.style.getPropertyValue('border') is '20px solid green'
PASS e.style.getPropertyPriority('border') is ''
PASS e.style.getPropertyValue('border') is '20px solid green'
PASS e.style.getPropertyPriority('border') is 'important'
PASS successfullyParsed is true

TEST COMPLETE
